requirements.txt
streamlit
pandas
pyserial